Changed File Structure for easier modification Removed long text blocks of flavor text to constant files.
