+++ src/constants/deathsMethods.js
@@ -1,0 +1,22 @@
+export const waysToDie = [
+        "was carried by the townsfolk and dropped into a nearby volcano.",
+        "was strapped to an ICBM by the town and launched out of the town. There was no parachute",
+        "was given a lethal dose of weed and perished instantly.",
+        "was picked up by a lord of the rings eagle and carried back to feed her chicks.",
+        "has been sent to start a single-person colony on Mars. We wish them luck",
+        "was bonked in the head with a 2x4 by a visiting gorilla.",
+        "was struck down by lightning on the spot. This could be an act of god.",
+        "was killed by a pretty nasty static shock from wearing a sweater that was just too fluffy.",
+        "was ran over by a herd of water buffalo migrating through the area.",
+        "ascended to a higher plane right there in the town, wow! Still dead though.",
+        "was immediately shot down by the town's oldest resident; Ms. Pancakes aged 93.",
+        "as a result their head exploded! Ew!",
+        "was pecked to death by a swarm of angry seagulls who thought they were a french fry.",
+        "was guillotined by the angry townsfolk!",
+        "was crushed by 10,000 bowling balls during the town's yearly bowling jubilee.",
+        "was sent on a submarine expedition to see the wreck of the Titanic.",
+        "while fleeing the townsfolk, quickly turned to run down what appeared to be a tunnel. However, the tunnel was actually just a painting of a tunnel. They crashed into the cliff at such velocity that their entire body was flattened into the thickness of a dinner plate.",
+        "was fed to the town's giant catfish: Skrimpo",
+        "was turned into a brundlefly by the town's local Jeff Goldbloom.",
+        "was sent down into the furniture mines, never to see the light of day again.",
+      ]
