+++ src/constants/edgyAdjectives.js
@@ -1,0 +1,52 @@
+export const abstractEdgyAdjectives = [
+  "Eldritch",
+  "Sonder",
+  "Liminal",
+  "Vexing",
+  "Oblivious",
+  "Hollow",
+  "Fractured",
+  "Gauche",
+  "Voidspun",
+  "Withering",
+  "Nihilistic",
+  "Haphazard",
+  "Feverish",
+  "Gnarled",
+  "Languid",
+  "Mordant",
+  "Pallid",
+  "Quicksand",
+  "Rancid",
+  "Spectral",
+  "Torpor",
+  "Umbral",
+  "Vexed",
+  "Wan",
+  "Xenolithic",
+  "Yawning",
+  "Zephyrous",
+  "Chthonic",
+  "Blighted",
+  "Crepuscular",
+  "Drifting",
+  "Effete",
+  "Fugue",
+  "Guttering",
+  "Howling",
+  "Inchoate",
+  "Jaded",
+  "Kaleidotic",
+  "Lurid",
+  "Miasmic",
+  "Nebulous",
+  "Ossified",
+  "Puckered",
+  "Queasy",
+  "Riven",
+  "Screaming",
+  "Threnodic",
+  "Unmoored",
+  "Vermicular",
+  "Wretched",
+];
