+++ src/constants/murderMethods.js
@@ -1,0 +1,23 @@
+export const murderMethods = [
+        "via blunt force trauma by someone wielding a pair of pool noodles. How persistant!",
+        "when the mafia hacked into their neuralink and ran SIGKILL.",
+        "by being pushed down the stairs from the third John Wick movie.",
+        "by a trained attack rhinocerous. Be on the lookout!",
+        "by a really gross fart.",
+        "using a commemorative Captain Sisko baseball bat.",
+        "by a swordfish ran right through the heart.",
+        "by stabbing, shooting, burning and poison at the same time.",
+        "using a wiffle bat.",
+        "outside of the town's public library, clubbed to death with their own books.",
+        "by someone who absolutely reeked of spaghetti sauce.",
+        "with a knife. Of course!",
+        "by a carefully planned evil Rube-Goldberg device hidden in the walls. The device failed catastrophically burning their house down.",
+        "with kindness? aww.",
+        "with pure telepathy!",
+        "by a massive black anvil, dropped from the sky.",
+        "by someone who had a bag of cheetos with them. There's orange dust everywhere.",
+        "by a remote operated flying lawnmower.",
+        'just mere seconds before they could get to their collection of "vintage" katanas',
+        "in glorious battle against an army of warriors from the opposing clans! Qa'Pla!",
+        "by someone dressed up as Robocop.",
+      ];
