+++ src/constants/wittyNouns.js
@@ -1,0 +1,63 @@
+export const wittyNouns = [
+  // Original 30
+  "Kazoo",
+  "Pajamas",
+  "Goblin",
+  "Mustache",
+  "Shenanigans",
+  "Fluff",
+  "Nonsense",
+  "Tomfoolery",
+  "Hullabaloo",
+  "Doohickey",
+  "Widget",
+  "Gadget",
+  "Whatchamacallit",
+  "Thingamajig",
+  "Malarkey",
+  "Hootenanny",
+  "Kerfuffle",
+  "Balderdash",
+  "Claptrap",
+  "Folderol",
+  "Snickerdoodle",
+  "Flibbertigibbet",
+  "Brouhaha",
+  "Skullduggery",
+  "Hullaballoon",
+  "Nincompoop",
+  "Skedaddle",
+  "Whippersnapper",
+  "Lollygag",
+  "Hobnob",
+  "Dinglehopper",
+  "Fandango",
+  "Gibberish",
+  "Hodgepodge",
+  "Hokum",
+  "Hullaballoo",
+  "Jiggerypokery",
+  "Klutz",
+  "Mumbojumbo",
+  "Poppycock",
+  "Ragamuffin",
+  "Rigmarole",
+  "Scallywag",
+  "Shenanigan",
+  "Skedaddler",
+  "Slapstick",
+  "Shenanigator",
+  "Thingummy",
+  "Widdershins",
+  "Yahoo",
+  "Zigzag",
+  "Blunderbuss",
+  "Cahoots",
+  "Doodad",
+  "Flapdoodle",
+  "Gobbledygook",
+  "Higgledypiggledy",
+  "Hullabuster",
+  "Knickerbocker",
+  "Lugubrious",
+];
